feature: Se modifica los componente de livewire

+++ app/Http/Livewire/ConfirmDeleteModel.php
@@ -2,19 +2,23 @@
 
 namespace App\Http\Livewire;
 
+use App\Models\User;
 use Livewire\Component;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Auth;
 
 class ConfirmDeleteModel extends Component
 {
+    public User $user;
     public Model $model;
     public $redirect = null;   // ruta a donde redirigir (opcional)
     public $message = 'Esta acción no se puede deshacer.';
     public $title = '¿Eliminar este registro?';
     public $showModal = false;
 
-    public function mount(Model $model, $redirect = null, $title = null, $message = null)
+    public function mount(User $user, Model $model, $redirect = null, $title = null, $message = null)
     {
+        $this->user = $user;
         $this->model = $model;
         $this->redirect = $redirect;
         if ($title) $this->title = $title;
@@ -23,6 +27,13 @@ class ConfirmDeleteModel extends Component
 
     public function delete()
     {
+        if(!$this->model && !$this->user ) return;
+
+        if(Auth::check() && Auth::user()->id != $this->user->id ) {
+            toastr()->addSuccess('Registro no eliminado, acción prohibida');
+            return;
+        }
+
         // Eliminar modelo
         $this->model->delete();
